+++ packages/common/src/BucketsConfig.ts
@@ -1,13 +1,21 @@
-'use strict';
+import { createErrorType } from '@cumulus/errors';
 
-const { createErrorType } = require('@cumulus/errors');
 const BucketsConfigError = createErrorType('BucketsConfigError');
+
+type BucketsConfigObject = {
+  [key: string]: {
+    name: string,
+    type: string
+  }
+};
 
 /**
  * Class representing cumulus bucket configuration.
  */
 class BucketsConfig {
-  constructor(bucketsJsonObject) {
+  private readonly buckets: BucketsConfigObject;
+
+  constructor(bucketsJsonObject: BucketsConfigObject) {
     this.buckets = bucketsJsonObject;
   }
 
@@ -16,7 +24,7 @@ class BucketsConfig {
    * @param {string} bucketName
    * @returns {string} desired bucket's key value.
    */
-  key(bucketName) {
+  key(bucketName: string) {
     return Object.keys(this.buckets)
       .find((bucketKey) => bucketName === this.buckets[bucketKey].name);
   }
@@ -26,7 +34,7 @@ class BucketsConfig {
    * @param {string} bucketName
    * @returns {string} matching bucket's type
    */
-  type(bucketName) {
+  type(bucketName: string) {
     const key = this.key(bucketName);
     if (!key) {
       throw new BucketsConfigError(`bucketName ${bucketName} not found in config ${JSON.stringify(this.buckets)}`);
@@ -39,7 +47,7 @@ class BucketsConfig {
    * @param {string} bucketName
    * @returns {Object} bucket object
    */
-  bucket(bucketName) {
+  bucket(bucketName: string) {
     const key = this.key(bucketName);
     if (!key) {
       throw new BucketsConfigError(`bucketName ${bucketName} not found in config ${JSON.stringify(this.buckets)}`);
@@ -52,7 +60,7 @@ class BucketsConfig {
    * @param {string} bucketName
    * @returns {boolean} truthyness of this bucket existing in the configuration
    */
-  exists(bucketName) {
+  exists(bucketName: string) {
     return this.key(bucketName) !== undefined;
   }
 
@@ -61,7 +69,7 @@ class BucketsConfig {
    * @param {string} configKey
    * @returns {boolean} truthyness of this key existing in the configuration
    */
-  keyExists(configKey) {
+  keyExists(configKey: string) {
     return Object.keys(this.buckets).includes(configKey);
   }
 
@@ -70,24 +78,19 @@ class BucketsConfig {
    * @param {string} configKey
    * @returns {string} name of bucket at key.
    */
-  nameByKey(configKey) {
+  nameByKey(configKey: string) {
     return this.buckets[configKey].name;
   }
 
   /**
    * return a list of configured buckets of desired type.
    *
-   * @param {string/Array} types - types of buckets to return
+   * @param {string} type - types of buckets to return
    * @returns {Array<Object>} - array of buckets that are of desired types
    */
-  bucketsOfType(types) {
-    const checkTypes = [].concat(types);
-    return Object.keys(this.buckets).map((key) => {
-      if (checkTypes.includes(this.buckets[key].type)) {
-        return this.buckets[key];
-      }
-      return undefined;
-    }).filter((valid) => valid);
+  bucketsOfType(type: string) {
+    return Object.values(this.buckets)
+      .filter((bucket) => bucket.type === type);
   }
 
   /** @returns {Array} list of private buckets */
@@ -116,5 +119,4 @@ class BucketsConfig {
   }
 }
 
-
-module.exports = BucketsConfig;
+export = BucketsConfig;